# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImportBatchRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  importKey: string;
  /**
   * @minLength 1
   * @maxLength 255
   */
  filename: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  mimeType: string;
  /** @minLength 1 */
  content: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  name?: string;
}

export interface ImportBatchResult {
  batchId: string;
  replayed: boolean;
  /** @minimum 1 */
  itemCount: number;
  /** @pattern ^[a-f0-9]{64}$ */
  checksum: string;
}

export interface BatchProgress {
  /** @minimum 0 */
  totalItemCount: number;
  /** @minimum 0 */
  completedCount: number;
  /** @minimum 0 */
  reviewCount: number;
  /** @minimum 0 */
  processedCount: number;
  /** @minimum 0 */
  pendingCount: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  percent: number;
  complete: boolean;
}

export interface Batch {
  id: string;
  name: string;
  status: string;
  progress: BatchProgress;
  [key: string]: unknown;
}

export interface IncludedItemsAnswer {
  selectedQuestionIds: string[];
  explicitlyNone: boolean;
}

export type ConditionCode = (typeof ConditionCode)[keyof typeof ConditionCode];

export const ConditionCode = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export type ItemDraftStatus =
  (typeof ItemDraftStatus)[keyof typeof ItemDraftStatus];

export const ItemDraftStatus = {
  new: "new",
  editing: "editing",
  restored: "restored",
  submitted: "submitted",
  needs_review: "needs_review",
} as const;

export type ItemDraftFieldValues = {
  [key: string]: string | number | boolean | null;
};

export interface ItemDraft {
  itemId: string;
  /** @minimum 1 */
  itemVersion: number;
  includedItems: IncludedItemsAnswer;
  conditionCode: ConditionCode | null;
  fieldValues: ItemDraftFieldValues;
  notes: string;
  /** @minLength 1 */
  employeeId: string;
  status: ItemDraftStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ListingItem {
  id: string;
  batchId: string;
  sku: string;
  status: string;
  /** @minimum 1 */
  version: number;
  draft?: ItemDraft | null;
  [key: string]: unknown;
}

export type DraftWriteResponseStatus =
  (typeof DraftWriteResponseStatus)[keyof typeof DraftWriteResponseStatus];

export const DraftWriteResponseStatus = {
  editing: "editing",
} as const;

export interface DraftWriteResponse {
  itemId: string;
  /** @minimum 1 */
  itemVersion: number;
  /** @minimum 1 */
  draftVersion: number;
  status: DraftWriteResponseStatus;
}

export interface SaveItemRequest {
  draft: ItemDraft;
}

export type NeedsReviewRequestReasonCode =
  (typeof NeedsReviewRequestReasonCode)[keyof typeof NeedsReviewRequestReasonCode];

export const NeedsReviewRequestReasonCode = {
  inventory_discrepancy: "inventory_discrepancy",
  item_damage: "item_damage",
  identity_uncertain: "identity_uncertain",
  missing_information: "missing_information",
  workflow_exception: "workflow_exception",
  other: "other",
} as const;

export type NeedsReviewRequestReason = {
  code: NeedsReviewRequestReasonCode;
  /** @maxLength 2000 */
  note?: string;
};

export interface NeedsReviewRequest {
  draft: ItemDraft;
  reason: NeedsReviewRequestReason;
}

export type ItemWriteResponseStatus =
  (typeof ItemWriteResponseStatus)[keyof typeof ItemWriteResponseStatus];

export const ItemWriteResponseStatus = {
  completed: "completed",
  needs_review: "needs_review",
} as const;

export interface ItemWriteResponse {
  itemId: string;
  /** @minimum 1 */
  itemVersion: number;
  status: ItemWriteResponseStatus;
  nextItemId: string | null;
  replayed: boolean;
}

export type ApiErrorCode = (typeof ApiErrorCode)[keyof typeof ApiErrorCode];

export const ApiErrorCode = {
  VALIDATION_ERROR: "VALIDATION_ERROR",
  INVALID_CSV: "INVALID_CSV",
  INVALID_ENCODING: "INVALID_ENCODING",
  INVALID_FILE_TYPE: "INVALID_FILE_TYPE",
  UPLOAD_TOO_LARGE: "UPLOAD_TOO_LARGE",
  ROW_LIMIT_EXCEEDED: "ROW_LIMIT_EXCEEDED",
  UNEXPECTED_HEADERS: "UNEXPECTED_HEADERS",
  UNSUPPORTED_SCHEMA_VERSION: "UNSUPPORTED_SCHEMA_VERSION",
  CONFLICT: "CONFLICT",
  SYSTEM_ERROR: "SYSTEM_ERROR",
  AI_PROCESSING_ERROR: "AI_PROCESSING_ERROR",
  UNAUTHORIZED: "UNAUTHORIZED",
  FORBIDDEN: "FORBIDDEN",
  NOT_FOUND: "NOT_FOUND",
  RATE_LIMITED: "RATE_LIMITED",
} as const;

export type ApiErrorFieldErrors = { [key: string]: string[] };

export interface ApiError {
  code: ApiErrorCode;
  message: string;
  correlationId: string;
  fieldErrors?: ApiErrorFieldErrors;
  currentVersion?: number;
}

/**
 * Structured API error
 */
export type ApiErrorResponse = ApiError;

export type IdempotencyKeyParameter = string;

export type ListBatches200 = {
  batches: Batch[];
};

export type ListBatchItems200 = {
  items: ListingItem[];
};

export type GetNextPendingItem200 = {
  item: ListingItem | null;
};