# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */
import type { ApiErrorCode } from "./apiErrorCode";
import type { ApiErrorFieldErrors } from "./apiErrorFieldErrors";

export interface ApiError {
  code: ApiErrorCode;
  message: string;
  correlationId: string;
  fieldErrors?: ApiErrorFieldErrors;
  currentVersion?: number;
}