# lib/api-zod/src/generated/types/draftWriteResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */
import type { DraftWriteResponseStatus } from "./draftWriteResponseStatus";

export interface DraftWriteResponse {
  itemId: string;
  /** @minimum 1 */
  itemVersion: number;
  /** @minimum 1 */
  draftVersion: number;
  status: DraftWriteResponseStatus;
}