# lib/api-zod/src/generated/types/importBatchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */

export interface ImportBatchResult {
  batchId: string;
  replayed: boolean;
  /** @minimum 1 */
  itemCount: number;
  /** @pattern ^[a-f0-9]{64}$ */
  checksum: string;
}