# lib/api-zod/src/generated/types/listingItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */
import type { ItemDraft } from "./itemDraft";

export interface ListingItem {
  id: string;
  batchId: string;
  sku: string;
  status: string;
  /** @minimum 1 */
  version: number;
  draft?: ItemDraft | null;
  [key: string]: unknown;
}