# lib/api-zod/src/generated/types/saveItemRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Executable controlled nonproduction integration API
 * OpenAPI spec version: 0.2.0
 */
import type { ItemDraft } from "./itemDraft";

export interface SaveItemRequest {
  draft: ItemDraft;
}